+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Feature to configure `VQT Menu Icon Resizer` during manual conversion when `VQT Avatar Converter Settings` is set to compress menu icons.
 - Cache to accelerate the texture generation process.
 - Support offline testing on Android in VQT Avatar Builder.
+- VRCQuestTools Settings to Unity project settings.
 
 ### Changed
 - Accelerated the texture generation process.